+++ test/is/acceptance_not.go
@@ -1,0 +1,5 @@
+// +build !acceptance
+
+package is
+
+const Acceptance = false