+++ test/is/integration.go
@@ -1,0 +1,5 @@
+// +build integration
+
+package is
+
+const Integration = true