+++ test/is/integration_test.go
@@ -1,0 +1,14 @@
+// +build integration
+
+package is_test
+
+import (
+	"testing"
+
+	"github.com/goph/fxt/test/is"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestIntegration(t *testing.T) {
+	assert.True(t, is.Integration)
+}