+++ test/is/unit.go
@@ -1,0 +1,5 @@
+// +build unit
+
+package is
+
+const Unit = true